+++ SDeMo/CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## `v1.6.1`
+
+- **added** tests for reliability curve and calibration
+- **improved** the performance of applying `PlattCalibration` [#522]
+- **added** support for the weights in PAVA [#521]
+- **improved** the thread safety of `Bagging` training
+- **improved** the performance of `DecisionTree`
+
 ## `v1.6.0`
 
 - **added** support for isotonic calibration